ConDec-428: Unify createDecisionKnowledgeElement and createDecisionKnowledgeElementAsChild methods into one method createDecisionKnowledgeElement
