+++ src/test/java/de/uhd/ifi/se/decision/management/jira/rest/knowledgerest/TestCreateDecisionKnowledgeElement.java
@@ -5,16 +5,16 @@ import static org.junit.Assert.assertEquals;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
 
-import de.uhd.ifi.se.decision.management.jira.ComponentGetter;
-import de.uhd.ifi.se.decision.management.jira.TestSetUpWithIssues;
-import de.uhd.ifi.se.decision.management.jira.mocks.MockTransactionTemplateWebhook;
-import net.java.ao.test.jdbc.Data;
-import net.java.ao.test.jdbc.NonTransactional;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import com.google.common.collect.ImmutableMap;
 
+import de.uhd.ifi.se.decision.management.jira.ComponentGetter;
+import de.uhd.ifi.se.decision.management.jira.TestSetUpWithIssues;
+import de.uhd.ifi.se.decision.management.jira.mocks.MockTransactionTemplateWebhook;
+import net.java.ao.test.jdbc.Data;
+import net.java.ao.test.jdbc.NonTransactional;
 import net.java.ao.test.junit.ActiveObjectsJUnitRunner;
 
 @RunWith(ActiveObjectsJUnitRunner.class)
@@ -26,7 +26,7 @@ public class TestCreateDecisionKnowledgeElement extends TestKnowledgeRestSetUp {
 	@Test
 	public void testRequestNullElementNull() {
 		assertEquals(Response.status(Status.BAD_REQUEST).entity(ImmutableMap.of("error", CREATION_ERROR)).build()
-				.getEntity(), knowledgeRest.createDecisionKnowledgeElement(null, null).getEntity());
+				.getEntity(), knowledgeRest.createDecisionKnowledgeElement(null, 0, null, null).getEntity());
 	}
 
 	@Test
@@ -34,7 +34,7 @@ public class TestCreateDecisionKnowledgeElement extends TestKnowledgeRestSetUp {
 		assertEquals(
 				Response.status(Status.BAD_REQUEST).entity(ImmutableMap.of("error", CREATION_ERROR)).build()
 						.getEntity(),
-				knowledgeRest.createDecisionKnowledgeElement(null, decisionKnowledgeElement).getEntity());
+				knowledgeRest.createDecisionKnowledgeElement(null, 0, null, decisionKnowledgeElement).getEntity());
 	}
 
 	@Test
@@ -42,16 +42,38 @@ public class TestCreateDecisionKnowledgeElement extends TestKnowledgeRestSetUp {
 		request.setAttribute("WithFails", false);
 		request.setAttribute("NoFails", true);
 		assertEquals(Response.status(Status.BAD_REQUEST).entity(ImmutableMap.of("error", CREATION_ERROR)).build()
-				.getEntity(), knowledgeRest.createDecisionKnowledgeElement(request, null).getEntity());
+				.getEntity(), knowledgeRest.createDecisionKnowledgeElement(request, 0, null, null).getEntity());
 	}
 
 	@Test
-    @NonTransactional
+	@NonTransactional
 	public void testRequestFilledElementFilled() {
 		request.setAttribute("WithFails", false);
 		request.setAttribute("NoFails", true);
 		ComponentGetter.setTransactionTemplate(new MockTransactionTemplateWebhook());
 		assertEquals(Status.OK.getStatusCode(),
-				knowledgeRest.createDecisionKnowledgeElement(request, decisionKnowledgeElement).getStatus());
+				knowledgeRest.createDecisionKnowledgeElement(request, 0, null, decisionKnowledgeElement).getStatus());
+	}
+
+	@Test
+	@NonTransactional
+	public void testRequestFilledElementFilledWithArgument() {
+		request.setAttribute("WithFails", false);
+		request.setAttribute("NoFails", true);
+		ComponentGetter.setTransactionTemplate(new MockTransactionTemplateWebhook());
+		decisionKnowledgeElement.setType("Pro-argument");
+		assertEquals(Status.OK.getStatusCode(),
+				knowledgeRest.createDecisionKnowledgeElement(request, 0, null, decisionKnowledgeElement).getStatus());
+	}
+
+	@Test
+	@NonTransactional
+	public void testRequestFilledElementFilledWithConArgument() {
+		request.setAttribute("WithFails", false);
+		request.setAttribute("NoFails", true);
+		ComponentGetter.setTransactionTemplate(new MockTransactionTemplateWebhook());
+		decisionKnowledgeElement.setType("Con-argument");
+		assertEquals(Status.OK.getStatusCode(),
+				knowledgeRest.createDecisionKnowledgeElement(request, 0, null, decisionKnowledgeElement).getStatus());
 	}
 }
